requirements.txt
numpy
Pillow
PyQt5
matplotlib
scikit-image